refactor: added rowUtils.tsx
