+++ static/onlyoffice-docs-confluence-cloud-custom-ui/src/components/ContentTree/index.tsx
@@ -18,17 +18,10 @@
 
 import React, { useEffect, useState } from "react";
 
-import { IconButton } from "@atlaskit/button/new";
-import DropdownMenu, {
-  DropdownItem,
-  DropdownItemGroup,
-} from "@atlaskit/dropdown-menu";
 import DynamicTable from "@atlaskit/dynamic-table";
 import { RowType } from "@atlaskit/dynamic-table/dist/types/types";
-import ShowMoreIcon from "@atlaskit/icon/core/show-more-horizontal";
-import { Box, Inline, Stack, xcss } from "@atlaskit/primitives";
-import { invoke, router } from "@forge/bridge";
-import bytes from "bytes";
+import { Stack } from "@atlaskit/primitives";
+import { invoke } from "@forge/bridge";
 
 import {
   findContent,
@@ -45,22 +38,7 @@ import { ContentTreePagination } from "./components/ContentTreePagination";
 import { ContentTreeToolbar } from "./components/ContentTreeToolbar";
 import { buildCreateRow } from "./utils/createRowUtils";
 import { getIconByContentType } from "./utils/iconUtils";
-
-const styles = {
-  searchIcon: xcss({
-    paddingLeft: "space.075",
-  }),
-  actionsContainer: xcss({
-    textAlign: "center",
-  }),
-  iconContainer: xcss({
-    display: "flex",
-    alignItems: "center",
-  }),
-  formFieldWraper: xcss({
-    marginBlockStart: "space.negative.100",
-  }),
-};
+import { buildContentTreeRows } from "./utils/rowUtils";
 
 export type ContentTreeProps = {
   space: {
@@ -199,7 +177,13 @@ export const ContentTree: React.FC<ContentTreeProps> = ({
 
     contentRequest
       .then((response) => {
-        const rows = buildContentTreeRows(response.results, contentType);
+        const rows = buildContentTreeRows(
+          currentParentId,
+          contentType,
+          response.results,
+          getDocumentType,
+          setCurrentParentId,
+        );
 
         setNavigationLinks(response._links);
         setRows(rows);
@@ -225,84 +209,6 @@ export const ContentTree: React.FC<ContentTreeProps> = ({
     }
 
     return "";
-  };
-
-  const buildContentTreeRows = (
-    entities: Content[],
-    contentType?: string,
-  ): RowType[] => {
-    const onClickOnTitle = (entity: Content) => {
-      if (entity.type !== "attachment") {
-        setCurrentParentId(entity.id);
-      }
-    };
-
-    return entities.map((entity) => ({
-      key: entity.id,
-      cells: [
-        {
-          key: "title",
-          content: (
-            <Box onClick={() => onClickOnTitle(entity)}>
-              <Inline space="space.075" alignBlock="center">
-                <Box xcss={styles.iconContainer}>
-                  {getIconByContentType(
-                    contentType || entity.type,
-                    getDocumentType(entity.title) || "",
-                  )}
-                </Box>
-                <Box>{entity.title}</Box>
-              </Inline>
-            </Box>
-          ),
-        },
-        {
-          key: "fileSize",
-          content: entity?.extensions?.fileSize
-            ? String(bytes(entity?.extensions?.fileSize))
-            : "",
-        },
-        {
-          key: "lastmodified",
-          content: entity.version.createdAt || entity.version.when,
-        },
-        {
-          key: "actions",
-          content: (
-            <Box xcss={styles.actionsContainer}>
-              <DropdownMenu<HTMLButtonElement>
-                trigger={({ triggerRef, ...props }) => (
-                  <IconButton
-                    {...props}
-                    icon={ShowMoreIcon}
-                    label="more"
-                    ref={triggerRef}
-                  />
-                )}
-                placement="bottom-end"
-                shouldRenderToParent
-              >
-                <DropdownItemGroup>
-                  <DropdownItem
-                    onClick={() =>
-                      router.open(
-                        `/forge-apps/a/f8a806c4-dbce-447e-9fc5-5edd102f13aa/e/52727433-e0ce-4dca-996e-7c0d911165cf/r/editor?pageId=${currentParentId}&attachmentId=${entity.id}`,
-                      )
-                    }
-                  >
-                    Edit
-                  </DropdownItem>
-                  <DropdownItem>Download</DropdownItem>
-                </DropdownItemGroup>
-                <DropdownItemGroup hasSeparator>
-                  <DropdownItem>Delete</DropdownItem>
-                </DropdownItemGroup>
-              </DropdownMenu>
-            </Box>
-          ),
-        },
-      ],
-    }));
   };
 
   const onChangeContentType = (selectedContentType: { value: string }) => {
